# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface OutfitRequest {
  prompt: string;
}

export interface OutfitJobResponse {
  jobId: string;
  status: string;
}

export interface PortalProduct {
  portal: string;
  portalColor: string;
  productName: string;
  brand: string;
  price: number;
  originalPrice: number;
  discountPercent: number;
  rating: number;
  reviewCount: number;
  productUrl: string;
  isBestBuy: boolean;
  bestBuyReason: string;
}

export interface OutfitItem {
  id: string;
  category: string;
  description: string;
  color: string;
  whyRecommended: string;
  portalProducts: PortalProduct[];
}

export interface Outfit {
  outfitTitle: string;
  styleDescription: string;
  occasion: string;
  targetProfile: string;
  totalBudgetMin: number;
  totalBudgetMax: number;
  items: OutfitItem[];
  outfitImage?: string | null;
}

export interface OutfitJob {
  jobId: string;
  status: string;
  prompt: string;
  outfits: Outfit[];
  error?: string | null;
  createdAt: string;
}