# lib/api-zod/src/generated/types/outfitItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PortalProduct } from "./portalProduct";

export interface OutfitItem {
  id: string;
  category: string;
  description: string;
  color: string;
  whyRecommended: string;
  portalProducts: PortalProduct[];
}